FIX: compile error
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@1051 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/dcm2image.cxx
@@ -439,7 +439,7 @@ VolumeList::WriteToArchive()
       for ( size_t i = 0; i < nVolumes; ++i )
 	{
 	std::ostringstream numberString;
-	numberString.width( 1 + static_cast<int>( log( nVolumes ) / log((float)10.0) ) );
+	numberString.width( 1 + static_cast<int>( log( nVolumes ) / log((double)10) ) );
 	numberString.fill( '0' );
 	numberString << std::right << 1+i;
